+++ NEWS.md
@@ -1,3 +1,10 @@
+MixedModels v3.3.0 Release Notes
+========================
+* HTML and LaTeX `show` methods for `MixedModel`, `BlockDescription`,
+  `LikelihoodRatioTest`, `OptSummary` and `VarCorr`. Note that the interface for
+  these is not yet completely stable. In particular, rounding behavior may
+  change. [#480]
+
 MixedModels v3.2.0 Release Notes
 ========================
 * Markdown `show` methods for `MixedModel`, `BlockDescription`,
@@ -137,3 +144,4 @@ Package dependencies
 [#447]: https://github.com/JuliaStats/MixedModels.jl/issues/447
 [#449]: https://github.com/JuliaStats/MixedModels.jl/issues/449
 [#474]: https://github.com/JuliaStats/MixedModels.jl/issues/474
+[#480]: https://github.com/JuliaStats/MixedModels.jl/issues/480